Rewrite logic of the game to work with UI
Select all the nodes in the DOM that are used to
play the games.

Define the images nodes that are appended to the
UI in coherence with the player and the computer
moves.

Prioritize saving snippets of code in functions,
which helps with the code readability.

The core of the code is in the event listener,
triggered by pressing any of the buttons.
This calls every function to behave a certain way
depending on the context.